requirements.txt
alabaster==0.7.13
autopep8==2.0.1
Babel==2.12.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
dgl==1.0.1+cu117
dglgo==0.0.2
dgllife==0.3.2
docutils==0.19
fair-esm==2.0.0
fonttools==4.38.0
future==0.18.3
hyperopt==0.2.7
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
isort==6.0.0b2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.1
littleutils==0.2.2
MarkupSafe==2.1.2
matplotlib==3.7.0
networkx==3.0
numpy==1.24.2
numpydoc==1.5.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
ogb==1.3.5
outdated==0.2.2
packaging==23.0
pandas==1.5.3
pandas-flavor==0.5.0
Pillow==9.4.0
psutil==5.9.4
py4j==0.10.9.7
pycodestyle==2.10.0
pydantic==1.10.5
pyg-lib==0.1.0+pt113cu117
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
rdkit-pypi==2022.9.5
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1
torch-cluster==1.6.0+pt113cu117
torch-geometric==2.2.0
torch-scatter==2.1.0+pt113cu117
torch-sparse==0.6.16+pt113cu117
torch-spline-conv==1.2.1+pt113cu117
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
xarray==2023.2.0
zipp==3.15.0
